refactor: extract profile launch gatherers to main service
Made-with: Cursor
